requirements.txt
discord.py[voice]>=2.3.2
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
Pillow>=10.0.0
yt-dlp>=2024.1.1
PyNaCl>=1.5.0
